requirements.txt
acres==0.3.0
ag2==0.8.6
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asyncer
attrs==25.3.0
autogen-agentchat==0.5.1
autogen-core==0.5.1
autogen-ext==0.5.1
azure-ai-inference==1.0.0b9
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
azure-search-documents==11.5.2
backoff==2.2.1
bidict==0.23.1
certifi==2025.1.31
cffi==1.17.1
chainlit
charset-normalizer==3.4.1
chevron==0.14.0
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dotenv==0.9.9
duckduckgo_search==8.0.0
etelemetry==0.3.1
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.6
jiter==0.9.0
jsonref==1.1.0
kiwisolver==1.4.8
Lazify==0.4.0
literalai==0.1.201
looseversion==1.3.0
lxml==5.3.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mcp==1.6.0
monotonic==1.6
msal==1.32.0
msal-extensions==1.3.1
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.4
openai==1.73.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.0
opentelemetry-instrumentation-anthropic==0.39.0
opentelemetry-instrumentation-bedrock==0.39.0
opentelemetry-instrumentation-chromadb==0.39.0
opentelemetry-instrumentation-cohere==0.39.0
opentelemetry-instrumentation-crewai==0.39.0
opentelemetry-instrumentation-google-generativeai==0.39.0
opentelemetry-instrumentation-groq==0.39.0
opentelemetry-instrumentation-haystack==0.39.0
opentelemetry-instrumentation-lancedb==0.39.0
opentelemetry-instrumentation-langchain==0.39.0
opentelemetry-instrumentation-llamaindex==0.39.0
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.39.0
opentelemetry-instrumentation-milvus==0.39.0
opentelemetry-instrumentation-mistralai==0.39.0
opentelemetry-instrumentation-ollama==0.39.0
opentelemetry-instrumentation-openai==0.39.0
opentelemetry-instrumentation-pinecone==0.39.0
opentelemetry-instrumentation-qdrant==0.39.0
opentelemetry-instrumentation-replicate==0.39.0
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.0
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.0
opentelemetry-instrumentation-transformers==0.39.0
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.0
opentelemetry-instrumentation-watsonx==0.39.0
opentelemetry-instrumentation-weaviate==0.39.0
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
posthog==3.24.1
primp==0.14.0
propcache==0.3.1
protobuf==5.29.4
prov==2.0.1
puremagic==1.28
pyautogen==0.8.6
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydot==3.0.4
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.2
#pywin32==310
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.41.3
syncer==2.0.3
tenacity==8.5.0
termcolor==3.0.1
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.39.0
traits==7.0.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uptrace==1.31.0
urllib3==2.4.0
uvicorn==0.34.1
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.19.0
zipp==3.21.0
chromadb
langchain-community
langchain-core
langchain-community
langchain-huggingface
docx2txt
pypdf
sentence-transformers
semantic-kernel
keybert
